+++ prp/parse/models/bracken.py
@@ -1,9 +1,12 @@
 """Bracken specific data models."""
 
-from pydantic import Field
+from typing import TypeAlias
+from pydantic import Field, TypeAdapter
+
+from prp.parse.core.registry import register_result_model
 
 from .base import BaseSpeciesPrediction
-from .enums import TaxLevel
+from .enums import TaxLevel, AnalysisSoftware, AnalysisType
 
 
 class BrackenSpeciesPrediction(BaseSpeciesPrediction):
@@ -13,3 +16,10 @@ class BrackenSpeciesPrediction(BaseSpeciesPrediction):
     kraken_assigned_reads: int = Field(..., alias="krakenAssignedReads")
     added_reads: int = Field(..., alias="addedReads")
     fraction_total_reads: float = Field(..., alias="fractionTotalReads")
+
+BrackenSpeciesPredictions: TypeAlias = list[BrackenSpeciesPrediction]
+
+register_result_model(
+    AnalysisSoftware.BRACKEN,
+    AnalysisType.SPECIES,
+)(TypeAdapter(BrackenSpeciesPredictions))