+++ prp/parse/parsers/tbprofiler.py
@@ -3,7 +3,9 @@
 from typing import Any, Sequence
 
 from prp.io.json import read_json
-from prp.parse.core.base import BaseParser, ParseImplOut, StreamOrPath
+from prp.io.types import StreamOrPath
+from prp.parse.core.base import BaseParser
+from prp.parse.models.base import ParseImplOut
 from prp.parse.core.envelope import run_as_envelope
 from prp.parse.core.registry import register_parser
 from prp.parse.exceptions import AbsentResultError
@@ -18,7 +20,7 @@ from prp.parse.models.enums import (
     VariantType,
 )
 from prp.parse.models.phenotype import TbProfilerVariant
-from prp.parse.models.typing import LineageInformation
+from prp.parse.models.typing import LineageInformation, LineageResults
 
 from .utils import get_db_version
 
@@ -155,7 +157,7 @@ def parse_drug_resistance_info(drugs: list[dict[str, str]]) -> list[PhenotypeInf
     return phenotypes
 
 
-def _to_lineage_result(pred: dict[str, Any]) -> list[LineageInformation]:
+def _to_lineage_result(pred: dict[str, Any]) -> LineageResults:
     """Transpose prediction result into a lineage object."""
 
     return [